requirements.txt
pysocks
urllib3[socks]
fastapi
uvicorn[standard]
qdrant-client
httpx
pyyaml
tiktoken
